[Kernel] Remove the CuTe DSL skinny GEMM
Now that the unquantized BF16 linear path dispatches through FlashInfer
mm_bf16 with backend="auto", the hand-tuned CuTe skinny GEMM and its
per-shape config tables are redundant. Delete the kernel and its residual
epilogue, and reduce the Kimi-K3 and GLM-5.2 low-latency tables to the
dsv3_fused_a_gemm entries they still need.

Shapes that were CuTe-only drop out of both tables and now fall through
to the default unquantized GEMM. The K3 residual path (apply_with_residual)
and the ParallelLMHead method go with them: the former had no production
caller and the latter no surviving head shape. GLM52_EH_PROJECTION was
CuTe-only and has no DISPATCH_DSV3_SHAPE(12288, 6144), so MTP's eh_proj
falls back to nn.Linear.

The ROCm skinny GEMMs (csrc/rocm/skinny_gemms.cu) and the latent MoE tail's
fused_add_multicast_skinny_gemm are unrelated kernels and are untouched.
