+++ vllm/models/deepseek_v32/nvidia/glm52_low_latency_gemm.py
@@ -12,67 +12,35 @@ from torch import nn
 
 import vllm.envs as envs
 from vllm import _custom_ops as ops
-from vllm.model_executor.kernels.linear.cute_dsl.skinny_gemm import (
-    SkinnyGemmConfig,
-    shape_dynamic_skinny_gemm,
-)
 from vllm.model_executor.layers.linear import (
     LinearBase,
     UnquantizedLinearMethod,
 )
 from vllm.platforms import current_platform
 
-Backend = Literal["cute", "dsv3_fused_a"]
-ResolvedCall = tuple[Backend, SkinnyGemmConfig | None]
+Backend = Literal["dsv3_fused_a"]
 
 
 @dataclass(frozen=True, slots=True)
 class GLM52ProjectionSpec:
     n: int
     k: int
-    cute_configs: tuple[tuple[int, SkinnyGemmConfig], ...]
     dsv3_tokens: frozenset[int] = frozenset()
 
-    def build_plan(self) -> dict[int, ResolvedCall]:
-        plan: dict[int, ResolvedCall] = {
-            num_tokens: ("cute", config) for num_tokens, config in self.cute_configs
-        }
-        plan.update(
-            (num_tokens, ("dsv3_fused_a", None)) for num_tokens in self.dsv3_tokens
-        )
-        return plan
+    def build_plan(self) -> dict[int, Backend]:
+        return {num_tokens: "dsv3_fused_a" for num_tokens in self.dsv3_tokens}
 
 
 GLM52_QKV_A_PROJECTION = GLM52ProjectionSpec(
     n=2624,
     k=6144,
-    cute_configs=(
-        (1, SkinnyGemmConfig(1, 128, 4, static_k=6144)),
-        (2, SkinnyGemmConfig(2, 128, 2)),
-    ),
     dsv3_tokens=frozenset(range(3, 17)),
 )
 
 GLM52_Q_B_PROJECTION = GLM52ProjectionSpec(
     n=2048,
     k=2048,
-    cute_configs=(
-        (1, SkinnyGemmConfig(1, 128, 4, static_k=2048)),
-        (2, SkinnyGemmConfig(2, 64, 2, k_unroll=2)),
-    ),
     dsv3_tokens=frozenset(range(3, 17)),
-)
-
-# The MTP eh_proj is a plain nn.Linear, so it gets its plan through
-# build_glm52_plan rather than a quant_method swap. cuBLAS wins from M=4.
-GLM52_EH_PROJECTION = GLM52ProjectionSpec(
-    n=6144,
-    k=12288,
-    cute_configs=(
-        (1, SkinnyGemmConfig(1, 256, 2, vector_width=4, static_k=12288)),
-        (2, SkinnyGemmConfig(2, 64, 2)),
-        (3, SkinnyGemmConfig(3, 64, 2)),
-    ),
 )
 
 GLM52_PROJECTIONS = {
@@ -80,7 +48,6 @@ GLM52_PROJECTIONS = {
     for spec in (
         GLM52_QKV_A_PROJECTION,
         GLM52_Q_B_PROJECTION,
-        GLM52_EH_PROJECTION,
     )
 }
 
@@ -108,22 +75,14 @@ def _runtime_ok(x: torch.Tensor, weight: torch.Tensor) -> bool:
 
 
 def run_glm52_plan(
-    plan: dict[int, ResolvedCall] | None,
+    plan: dict[int, Backend] | None,
     x: torch.Tensor,
     weight: torch.Tensor,
 ) -> torch.Tensor | None:
     if plan is None or not _runtime_ok(x, weight):
         return None
-    entry = plan.get(x.shape[0])
-    if entry is None:
+    if plan.get(x.shape[0]) is None:
         return None
-
-    backend, config = entry
-    if backend == "cute":
-        if not shape_dynamic_skinny_gemm.is_available():
-            return None
-        return shape_dynamic_skinny_gemm(x, weight, config)
-
     if not hasattr(torch.ops._C, "dsv3_fused_a_gemm"):
         return None
     output = torch.empty(
@@ -135,28 +94,20 @@ def run_glm52_plan(
     return output
 
 
-def _request_warmup(dtype: torch.dtype, configs: set[SkinnyGemmConfig]) -> None:
-    if configs and shape_dynamic_skinny_gemm.is_available():
-        shape_dynamic_skinny_gemm.request_warmup_configs(dtype, configs)
-
-
 def build_glm52_plan(
     weight: torch.Tensor | None, dtype: torch.dtype
-) -> dict[int, ResolvedCall] | None:
+) -> dict[int, Backend] | None:
     """Plan for a weight the walk below cannot reach (a plain ``nn.Linear``)."""
     if dtype != torch.bfloat16 or not _is_sm103():
         return None
     if weight is None or weight.dim() != 2 or weight.dtype != torch.bfloat16:
         return None
     spec = GLM52_PROJECTIONS.get(tuple(weight.shape))
-    if spec is None:
-        return None
-    _request_warmup(dtype, {config for _, config in spec.cute_configs})
-    return spec.build_plan()
+    return spec.build_plan() if spec is not None else None
 
 
 class GLM52LowLatencyLinearMethod(UnquantizedLinearMethod):
-    def __init__(self, plan: dict[int, ResolvedCall]) -> None:
+    def __init__(self, plan: dict[int, Backend]) -> None:
         super().__init__()
         self._plan = plan
 
@@ -180,7 +131,6 @@ def enable_glm52_low_latency_gemm(
     if dtype != torch.bfloat16 or not _is_sm103():
         return
 
-    warmup_configs: set[SkinnyGemmConfig] = set()
     for child in module.modules():
         if (
             not isinstance(child, LinearBase)
@@ -194,6 +144,3 @@ def enable_glm52_low_latency_gemm(
         if spec is None:
             continue
         child.quant_method = GLM52LowLatencyLinearMethod(spec.build_plan())
-        warmup_configs.update(config for _, config in spec.cute_configs)
-
-    _request_warmup(dtype, warmup_configs)
